point-rewards-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
python-dotenv
Werkzeug
pycryptodome
pandas
openpyxl